+++ history.md
@@ -12,7 +12,8 @@
 
 ### New Features
 
-1. Schedules can now be based on days of the week.
+1. Schedules can now be based on days of the week. [[#32](https://github.com/chadwpalm/PrerollPlus/discussions/32)]
+2. Added ability to include subdirectories when using a directory as a file source for buckets. [[#33](https://github.com/chadwpalm/PrerollPlus/issues/33)]
 
 ### Fixes